Stream V: 38 P1 items — grad-flow, parity, precision, GIL, unchecked async
audit-4 P1.  Lands every Stream V item.

Autograd (V.1–V.11):
* V.1 vmap_loop_and_stack rebuilt via autograd::unsqueeze + autograd::cat —
  grad_fn now survives create_graph under vmap.
* V.2 execute_multi gained create_graph parameter + dual-path block +
  stub-disconnection tracking matching the single-root path.
* V.3 execute_multi hoists accum_guard_multi ScopeGuard to immediately
  after std::move(grad_accumulators_) so accumulate_grad throws can't
  lose outer state.
* V.4 single-root leaf accumulation now uses promote_types(existing,
  incoming) instead of downcasting to leaf dtype; multi-root already
  used promote_types.
* V.5 Function::save_for_backward records source device per-tensor in
  a new offloaded_devices_ vector; reload sends each back to its own
  origin instead of the first device.
* V.6 jvp_leaky_relu / jvp_elu / jvp_softplus helper sigs and the
  three adapters widened float → double (R.2 holdouts).
* V.7 LinalgVectorNormBackward drops static_cast<float>(ord_-1.0);
  uses tenzor::pow(Tensor, double).
* V.8 composed_attention_backward (Tensor form), try_fused_or_compose_
  backward, try_flex_backward_or_throw scale widened float → double
  (R.6 holdouts).
* V.9 NarrowBackward / RollBackward / DiagBackward / TriuBackward /
  TrilBackward gained saved_attributes() + name() overrides; vmap
  registry replaced raw passthrough_rule with dim_shifted_passthrough
  / loop-and-stack as appropriate.
* V.10 new autograd::chunk and autograd::split Variable-level
  decompositions in ops.hpp / ops.cpp via repeated autograd::slice;
  each output carries a real SliceBackward.
* V.11 GatherBackward / ScatterBackward / ScatterAddBackward (both
  backward and backward_with_variables) normalise dim at entry;
  saved_attributes() emits the normalised non-negative dim.

Backend kernels (V.12–V.20):
* V.12 CPU Unfold/Fold throws on asymmetric per-axis attrs (matches
  CUDA/ROCm/OneAPI contract).
* V.13 Vulkan dispatchCol2Im reads per-axis attrs, throws on
  asymmetric kernel/stride/padding/dilation.
* V.14 ROCm chunk_kernel normalises negative dim with range check.
* V.15 ROCm bincount wraps both hipMemcpyAsync, hipStreamSynchronize,
  and hipMemsetAsync in HIP_CHECK.
* V.16 ROCm flash_attention_f64 plumbs hipStream_t through
  fused_attention_hip_f64 / flash_attention_backward_hip_f64 /
  create_hip_zeros_f64; all 14 hipLaunchKernelGGL sites now use the
  supplied stream; workspace zero is HIP_CHECK'd.
* V.17 CUDA cudaMemcpyAsync sweep — TENZOR_CUDA_CHECK on every
  unguarded site across advanced.cu (multinomial, STFT/ISTFT,
  bincount, isin, minmax, unique), math.cu (conjugate/real),
  rnn_sequence.cu (LSTM/GRU per-step, multi-layer, bidirectional),
  reduction.cu (n==1 short-circuits), transform.cu (slice loop).
* V.18 CUDA fused_ops.cu Philox seed synchronous cudaMemcpy
  wrapped in TENZOR_CUDA_CHECK (Int64 + Int32 paths).
* V.19 CUDA advanced_index_{gather,put} now use a cudaMallocHost
  pinned staging buffer for the index pointer arrays (was pageable
  std::vector blocking the host).  HtoD copy + cudaFreeHost cleanup
  both checked.
* V.20 covered by V.17 transform.cu slice-loop fix.

nn / optim / distributed (V.21–V.30):
* V.21 SWA::apply_to uses dst.zero_(); add_(dst, src_cast) instead of
  tensor reassignment; preserves Storage handle (R.18 invariant).
* V.22 FSDP2 keeps two slots (unsharded_dst_ / sharded_dst_) so the
  in-place zero_+add_ path runs every cycle instead of only round 1.
  Added unordered_map members to the FSDP2 class.
* V.23 nn::Checkpoint::write_checkpoint gates the RNG section emit on
  checkpoint.version >= 2 (matches the read-side gate).  Downgrade
  workflows now produce v1-compatible bytes.
* V.24 LayerNorm backward_with_variables reads saved[3] only inside
  if (elementwise_affine_); slot-3 contract documented at method head.
* V.26 SWA n_averaged_ = 0 at ctor (was 1); update_parameters
  special-cases n_averaged_ == 0 to install the first sample as a
  plain copy.  Cross-PyTorch checkpoint parity restored.
* V.28 Attention key_padding_mask: Bool / integer dtypes widen to
  Float32 before threshold/neg_inf/zero/where construction.  The
  previous full(0.5, Bool) truncating to 0 path is fixed.
* V.29 LazyLinear honours pre-materialisation dtype via new
  requested_dtype_ / requested_device_ members; Module::to(DType) and
  Module::to(Device) are now virtual.  Standard model.to(BFloat16);
  model(input) recipe produces BF16 params.

Python frontend (V.31–V.41):
* V.31 nn.Parameter _is_parameter flag now consulted by __setattr__
  before requires_grad fallback.
* V.32 autograd.grad captures + restores caller .grad and retain_grad
  state.  New Variable::set_retain_grad(bool) accessor.
* V.33 DataLoader probes new cuda_is_initialized() binding; falls
  back to spawn/forkserver when CUDA is active.
* V.34 PyInferenceModeContext + PySetGradEnabledContext LIFO-safe via
  vector<bool> / vector<shared_ptr<...>>.
* V.35 Tensor tuple-index accepts None / Ellipsis / bool mask via
  extended TupleEntry { Int, Slice, NewAxis, Ellipsis, BoolMask };
  dim_cursor tracks insert/expand/consume semantics.
* V.36 DDP.sync_comm / auto_sync_gradients / reset_buckets, FSDP.
  finalize_backward / summon_full_params / release_full_params, and
  rpc.rpc_sync gained py::call_guard<gil_scoped_release>().
* V.37 every compute-bearing .def in bindings_nested.cpp /
  bindings_vision_detection.cpp / bindings_quantization.cpp gained
  GIL release.
* V.38 audit_tolerances.py moved scripts/ → tools/; ctest entry and
  cross-refs updated.
* V.39 tenzor.distributions made lazy on scipy via
  _require_scipy_special / _require_scipy_stats; 15 distribution
  files converted.  tenzor.distributions now imports without scipy
  installed.
* V.40 check_param_group_serialization splits ModuleNotFoundError
  (exit 2) from other exceptions (exit 1); top-level try/except
  ensures non-zero exit on unhandled.
* V.41 BACKENDS multi now appends literal 'multi' label alongside
  the 5 backends; ctest -L multi works.  test_simd_ops
  DISABLED_AddPerformance converted to AddPerformance with
  EXPECT_LT ceiling + benchmarks/CMakeLists.txt wiring.

Verified: full ninja build clean in single -j1 shell.  No test runs
this commit per the only-narrowest-selector rule combined with the
parallel-edit churn from V.1-V.9 in engine/function/vmap.

+++ src/autograd/function_attention.cpp
@@ -104,10 +104,15 @@ auto composed_attention_backward(const Tensor& dO,
                                  const Tensor& Q,
                                  const Tensor& K,
                                  const Tensor& V,
-                                 float scale,
+                                 double scale,
                                  bool causal,
-                                 float dropout_p = 0.0f,
+                                 double dropout_p = 0.0,
                                  uint32_t rng_seed = 0u) -> std::vector<Tensor> {
+    // V.8: double scale end-to-end; audit-3 R.6 had widened only the
+    // Variable form. dropout_p is also widened to keep parity with the
+    // Variable path; the only downstream effect is the threshold used
+    // when constructing the dropout mask (precision-irrelevant for the
+    // mask itself but consistent with the rest of the signature).
     // Q, K, V, dO all have shape [B, H, S, D] or [B*H, S, D]. We work on
     // 3D for bmm convenience by collapsing leading dims if 4D. shape() returns
     // a span; convert to vector for the reshape API.
@@ -229,11 +234,14 @@ auto try_fused_or_compose_backward(const Tensor& dO,
                                    const Tensor& V,
                                    const Tensor& O,
                                    const Tensor& L,
-                                   float scale,
+                                   double scale,
                                    bool causal,
-                                   float dropout_p,
+                                   double dropout_p,
                                    const Tensor& philox_seed,
                                    const Tensor& philox_offset) -> std::vector<Tensor> {
+    // V.8: double scale/dropout_p; the callers (FlashAttentionBackward,
+    // FusedAttentionBackward) already hold `double scale_` and were
+    // narrowing through the float parameter at every call.
     // Read seed (if any) for the composed-fallback dropout replay.
     uint32_t seed_for_replay = 0u;
     if (philox_seed.is_valid() && philox_seed.numel() > 0) {
@@ -278,9 +286,10 @@ auto try_flex_backward_or_throw(const Tensor& dO,
                                 const Tensor& V,
                                 const Tensor& O,
                                 const Tensor& L,
-                                float scale,
+                                double scale,
                                 int64_t score_mod_id,
                                 const Tensor& block_mask) -> std::vector<Tensor> {
+    // V.8: double scale; the FlexAttentionBackward holds `double scale_`.
     std::vector<Tensor> bwd_inputs = {dO, Q, K, V, O};
     if (L.is_valid() && L.shape().size() > 0) bwd_inputs.push_back(L);
     if (block_mask.is_valid() && block_mask.shape().size() > 0) bwd_inputs.push_back(block_mask);
@@ -399,8 +408,11 @@ auto composed_attention_backward_variable(const Variable& dO,
                                           const Variable& Q,
                                           const Variable& K,
                                           const Variable& V,
-                                          float scale,
+                                          double scale,
                                           bool causal) -> std::vector<Variable> {
+    // V.8: double scale (the Variable form was the Tensor form's twin;
+    // both now take `double` end-to-end, matching the `double scale_`
+    // member on FlashAttentionBackward/FusedAttentionBackward/FlexAttentionBackward).
     auto Q_shape_span = Q.tensor().shape();
     std::vector<int64_t> orig_shape(Q_shape_span.begin(), Q_shape_span.end());
     std::vector<int64_t> K_shape(K.tensor().shape().begin(), K.tensor().shape().end());
